Setup
Created the basic backend objects and added all the necessary properties. I mapped them to each other with either one to many/ many to one / or many to many. Was able to link my database to have it up and running so that I can start working on the Controller.